infra/requirements.txt
aws-cdk-lib>=2.75.0
constructs>=10.0.0
